feat: adding input validations

+++ scripts/mainmenu.sh
@@ -5,29 +5,51 @@ PS3="Choose an option:"
 select choice in "Create Database" "List Databases" "Connect database" "Drop database" "Exit" 
 do
     case "$REPLY" in
-        1)
+        1)  #Create Database
             read -p "Enter database name: " db
+	    #-z checks if string length is zero
+	    if [ -z "$db" ]; then
+                echo "Database name cannot be empty"
+                continue
+            fi
             #Hides “File exists” error and echo created if succeed or exist if failed to create
 	    mkdir ../Databases/"$db" 2>/dev/null && echo "Database created" || echo "Database already exists"
             ;;
-        2)
+        2)  #List Databases
             echo "Databases:"
             ls -d ../Databases/* 2>/dev/null || echo "No databases found"
             ;;
-        3)
+        3)  #Connect database
             read -p "Enter database name to connect: " db
-            if [ -d ../Databases/"$db" ]; then
+            if [ -z "$db" ]; then
+                echo "Database name cannot be empty"
+                continue
+            fi
+	    if [ -d ../Databases/"$db" ]; then
                 cd ../Databases/"$db" || exit
                 echo "Connected to $db"
             else
                 echo "Database not found"
             fi
             ;;
-        4)
+        4)  #Drop database
             read -p "Enter database name to drop: " db
-            rm -r ../Databases/"$db" 2>/dev/null && echo "Database dropped" || echo "Database not found"
-            ;;
-        5)
+            if [ -z "$db" ]; then
+                echo "Database name cannot be empty"
+                continue
+            fi
+
+	    read -p "Are you sure you want to drop '$db'? (y/n): " confirm
+            case "$confirm" in
+       	    y|Y)
+           	 rm -r ../Databases/"$db" 2>/dev/null && echo "Database dropped" || echo "Database not found"
+            	 ;;
+            *)
+            	 echo "Drop cancelled"
+            	 ;;
+    	    esac
+	    ;;
+        5)  #Exit
             echo "Exiting..."
             exit
             ;;
